requirements.txt
redis[hiredis]
python-dotenv

